[Bronze V] Title: Multiply, Time: 48 ms, Memory: 5496 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.cs
@@ -2,6 +2,6 @@ using var sr = new StreamReader(new BufferedStream(Console.OpenStandardInput()))
 using var sw = new StreamWriter(new BufferedStream(Console.OpenStandardOutput()));
 
 sr.ReadLine();
-int a = int.Parse(sr.ReadLine());
-int b = int.Parse(sr.ReadLine());
+ulong a = ulong.Parse(sr.ReadLine());
+ulong b = ulong.Parse(sr.ReadLine());
 sw.WriteLine(a*b);
